+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.18.0
 - State: Multiple create action in GUI
+- JSON properties: define sub fields
 
 # 0.17.0
 - Adapter API: bulk add